added: icon, input-line, ui-button
icon imports fontawesome 4.7 and uses an icon property with unicode strings.
ui-button is a button element with different types available, can display Icon.
input-line is a textinput that can also render a button at the end
theme struct is a structure that is suposed to hold all theme data and is passed down to most children so there will be one place to change the look
custom-widgets reexports all custom components for cleaner imports in other modules